+++ 2ndUKF/src/ukf.h
@@ -10,22 +10,22 @@ public:
     ukf( int state_size , int measurement_size);
     ~ukf();
 
-    double L;
-    double dt;
+    float L;
+    float dt;
     //vector size
     int x_size;
     int y_size;
     int x_sigmavector_size;
     int y_sigmavector_size;
     virtual Eigen::MatrixXd  dynamics(Eigen::MatrixXd  sigma_state);
-    Eigen::MatrixXd rotate(double roll , double yaw , double pitch);
+    Eigen::MatrixXd rotate(float roll , float yaw , float pitch);
 
     void set_process_noise(Eigen::MatrixXd matrix);
     void set_measurement_noise(Eigen::MatrixXd matrix);
     void set_covariance_matrix(Eigen::MatrixXd matrix);
 
     void set_measurement_matrix(Eigen::MatrixXd matrix);
-    void set_parameter(double alpha,double beta , double lambda , double kappa);
+    void set_parameter(float alpha,float beta , float lambda , float kappa);
 
     void predict();
     void correct(Eigen::VectorXd measure);
@@ -38,10 +38,10 @@ public:
     Eigen::VectorXd x_hat; //x mean
     Eigen::VectorXd y_hat; //y mean
 
-    double alpha ;
-    double kappa ;
-    double beta ;
-    double lambda ;
+    float alpha ;
+    float kappa ;
+    float beta ;
+    float lambda ;
 private:
 
     Eigen::VectorXd w_c ; //weight c
@@ -64,8 +64,8 @@ private:
 
     Eigen::MatrixXd Kalman_gain ;
 
-    // double y_hat = 0.0;
-    // double x_hat = 0.0;
+    // float y_hat = 0.0;
+    // float x_hat = 0.0;
 
 };
